Use the default Devise validations
I can only think of bad things that would come of rolling our own
password validations. Use the ones that come with Devise instead.

Note the deleted spec: This _does_ break the ability to create a user
without a password, but I can't think of why one would want to do that.
That user can't log in, or register, or anything, so an admin would have
to set their password before their login is of any use.