Don't use the system libxml2
Who knows why this was added (it appears to have been copied over from
another app). But, it hsan't _done_ anything as long as nokogiri has
had precompiled gems available.

After #363, though, Bundler is now compiling the gem natively. And, it
turns out the the RHEL-provided version of libxml2 is too old. Luckily,
nokogiri ships with a vendored copy of that C library. We just have to
not not use it.